form field refactoring
